requirements.txt
amqp==5.1.1
aniso8601==7.0.0
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
dateparser==1.1.4
defusedxml==0.7.1
dj-database-url==1.2.0
Django==4.1.4
django-appconf==1.0.5
django-environ==0.9.0
django-formtools==2.4
django-otp==1.1.4
django-user-accounts==3.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
lxml==4.9.2
MarkupSafe==2.1.1
oauthlib==3.2.2
phonenumbers==8.13.2
Pillow==9.3.0
promise==2.3
prompt-toolkit==3.0.36
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
Rx==1.6.3
six==1.16.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.45
sqlparse==0.4.3
twilio==7.16.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
